requirements.txt
annotated-types==0.7.0
boto3==1.42.49
botocore==1.42.49
dynaconf==3.2.12
iniconfig==2.3.0
jmespath==1.1.0
packaging==26.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
s3transfer==0.16.0
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
python-dotenv==1.2.1
requests==2.31.0